Correction migrations commande, article

+++ database/migrations/2021_06_30_091016_create_articles_table.php
@@ -18,7 +18,7 @@ class CreateArticlesTable extends Migration
             $table->string('photo');
             $table->string('libelle');
             $table->string('categorie');
-            $table->float('prix');
+            $table->double('prix');
 
         });
     }
